backend/routes/transcribe.py
"""
Voice transcription route using speechrecognition
"""

import asyncio
import base64
import io
import wave
from typing import Dict, Any

# Try to import speech recognition, fallback if not available
try:
    import speech_recognition as sr
    SPEECH_RECOGNITION_AVAILABLE = True
except ImportError:
    SPEECH_RECOGNITION_AVAILABLE = False

async def process_transcription(audio_data: str, language: str) -> Dict[str, Any]:
    """
    Process audio transcription using speech recognition
    
    Args:
        audio_data: Base64 encoded audio data
        language: Language code for transcription
        
    Returns:
        Dict containing transcription result
    """
    try:
        # Decode base64 audio data
        audio_bytes = base64.b64decode(audio_data)
        
        # Convert to WAV format if needed
        audio_wav = await convert_to_wav(audio_bytes)
        
        # Check if speech recognition is available
        if not SPEECH_RECOGNITION_AVAILABLE:
            text = get_mock_transcription(language)
        else:
            # Initialize recognizer
            recognizer = sr.Recognizer()
            
            # Set language for recognition
            language_code = get_language_code(language)
            
            # Convert audio bytes to AudioData
            audio_data_obj = sr.AudioData(
                audio_wav, 
                sample_rate=16000, 
                sample_width=2
            )
            
            # Perform speech recognition
            try:
                text = recognizer.recognize_google(
                    audio_data_obj, 
                    language=language_code
                )
            except AttributeError:
                # Fallback if speech recognition method is not available
                text = get_mock_transcription(language)
        
        return {
            "success": True,
            "text": text,
            "language": language,
            "confidence": 0.85  # Mock confidence score
        }
        
    except sr.UnknownValueError:
        return {
            "success": False,
            "error": "Could not understand audio",
            "language": language
        }
    except sr.RequestError as e:
        return {
            "success": False,
            "error": f"Speech recognition service error: {str(e)}",
            "language": language
        }
    except Exception as e:
        return {
            "success": False,
            "error": f"Transcription error: {str(e)}",
            "language": language
        }

async def convert_to_wav(audio_bytes: bytes) -> bytes:
    """
    Convert audio bytes to WAV format
    
    Args:
        audio_bytes: Raw audio bytes
        
    Returns:
        WAV format audio bytes
    """
    try:
        # For demo purposes, return as-is
        # In production, you would convert to WAV format
        return audio_bytes
    except Exception as e:
        raise Exception(f"Audio conversion error: {str(e)}")

def get_language_code(language: str) -> str:
    """
    Get speech recognition language code
    
    Args:
        language: Language identifier
        
    Returns:
        Language code for speech recognition
    """
    language_codes = {
        'en': 'en-US',
        'hi': 'hi-IN',
        'ta': 'ta-IN',
        'te': 'te-IN',
        'kn': 'kn-IN',
        'ml': 'ml-IN',
        'gu': 'gu-IN',
        'mr': 'mr-IN',
        'bn': 'bn-IN',
        'or': 'or-IN',
        'pa': 'pa-IN'
    }
    
    return language_codes.get(language, 'en-US')

def validate_audio_format(audio_bytes: bytes) -> bool:
    """
    Validate audio format
    
    Args:
        audio_bytes: Audio data bytes
        
    Returns:
        True if valid format
    """
    # Check if it's a valid audio format
    # This is a simplified check
    return len(audio_bytes) > 1000  # Minimum size check

def get_audio_duration(audio_bytes: bytes) -> float:
    """
    Get audio duration in seconds
    
    Args:
        audio_bytes: Audio data bytes
        
    Returns:
        Duration in seconds
    """
    try:
        # For demo, return estimated duration
        # In production, use proper audio analysis
        return len(audio_bytes) / 16000  # Rough estimation
    except Exception:
        return 0.0

def get_mock_transcription(language: str) -> str:
    """
    Get mock transcription for demo purposes
    
    Args:
        language: Language code
        
    Returns:
        Mock transcribed text
    """
    mock_texts = {
        'en': "Fresh tomatoes from my farm, organic and healthy",
        'hi': "मेरे खेत से ताजे टमाटर, जैविक और स्वस्थ",
        'ta': "என் பண்ணையில் இருந்து புதிய தக்காளிகள், கரிம மற்றும் ஆரோக்கியமான",
        'te': "నా పొలం నుండి తాజా టమాటాలు, సేంద్రీయ మరియు ఆరోగ్యకరమైన",
        'kn': "ನನ್ನ ಕೃಷಿ ಭೂಮಿಯಿಂದ ತಾಜಾ ಟೊಮೇಟೊಗಳು, ಸಾವಯವ ಮತ್ತು ಆರೋಗ್ಯಕರ",
        'ml': "എന്റെ കൃഷിഭൂമിയിൽ നിന്നുള്ള പുതിയ തക്കാളികൾ, ജൈവവളവും ആരോഗ്യകരവും",
        'gu': "મારા ખેતરમાંથી તાજા ટામેટા, સેન્દ્રિય અને સ્વસ્થ",
        'mr': "माझ्या शेतातून ताजे टोमॅटो, सेंद्रिय आणि निरोगी",
        'bn': "আমার খামার থেকে তাজা টমেটো, জৈব এবং স্বাস্থ্যকর",
        'or': "ମୋ କ୍ଷେତରୁ ତାଜା ଟମାଟୋ, ସଜୀବ ଏବଂ ସୁସ୍ଥ",
        'pa': "ਮੇਰੇ ਖੇਤ ਤੋਂ ਤਾਜ਼ੇ ਟਮਾਟਰ, ਜੈਵਿਕ ਅਤੇ ਸਿਹਤਮੰਦ"
    }
    
    return mock_texts.get(language, mock_texts['en']) 