requirements.txt
fastapi==0.115.0
uvicorn==0.32.1
requests==2.31.0
pydantic==2.10.4
python-multipart==0.0.9
supabase==2.10.0
google-generativeai==0.8.3
python-dotenv==1.0.1 